% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLabels.R
\name{createLabels}
\alias{createLabels}
\title{Build labeldt for mlho_features}
\usage{
createLabels(days.long, rulesPhen, PatientObservations)
}
\arguments{
\item{days.long}{"long" stage threshold}

\item{rulesPhen}{table containing the rules for a specific phenotype}

\item{PatientObservations}{2.1 PatientObservations table}
}
\value{
labeldt
}
\description{
Build a table containing for each patient a label that specify whether the patient
matches the phenotype for the first time \code{days.long} after positive test
}
