% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{filter_data}
\alias{filter_data}
\title{Filter data according to type}
\usage{
filter_data(PatientObservations, PatientSummary, data_type)
}
\arguments{
\item{PatientObservations}{4CE table}

\item{PatientSummary}{4CE table}

\item{data_type}{string with the type of 4CE data}
}
\value{
filtered 4CE tables
}
\description{
Filter data according to type
}
