% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPublicSummaryRepositoryUrl.R
\name{getPublicSummaryRepositoryUrl}
\alias{getPublicSummaryRepositoryUrl}
\title{Returns the GitHub URL of the Public Summary Repository for this project}
\usage{
getPublicSummaryRepositoryUrl()
}
\description{
Returns the GitHub URL of the Public Summary Repository for this project
}
\keyword{4CE}
