% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlho_features.R
\name{mlho_features}
\alias{mlho_features}
\title{Apply MLHO to extract features}
\usage{
mlho_features(
  PatientObservations,
  PatientSummary,
  rules,
  days.long,
  days.pre = -14,
  phenotype,
  prediction = FALSE,
  test.sample = 25,
  MSMR.sparsity = 0.005,
  MSMR.topn = 200,
  multicore = FALSE,
  iteration = 5,
  score.threshold = 80,
  classifier
)
}
\arguments{
\item{PatientObservations}{2.1 PatientObservations table}

\item{PatientSummary}{2.1 PatientSummary table}

\item{rules}{table containing the phenotype definitions}

\item{days.long}{"long" stage threshold}

\item{days.pre}{"pre" stage threshold}

\item{phenotype}{phenotype label}

\item{prediction}{do you want to make prediction or phenotyping?}

\item{test.sample}{mlho.it parameter. Percentage for testing}

\item{MSMR.sparsity}{MSMR.lite parameter}

\item{MSMR.topn}{MSMR.lite parameter}

\item{multicore}{do you want to parallelize the process?}

\item{iteration}{mlho.it parameter. Number of iterations}

\item{score.threshold}{MLHO score threshold}

\item{classifier}{mlearn parameter. Classification algorithm}
}
\value{
MLHO features and their prevalence
}
\description{
Apply MLHO to extract features
}
