% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasc_bucks.R, R/pascbuckets.R
\name{pascbuckets}
\alias{pascbuckets}
\title{analyzing buckets of MLHO and core features}
\usage{
pascbuckets(
  data_dir,
  output_dir,
  siteid,
  long.thres,
  data_type,
  pascphens,
  chart_review = FALSE
)

pascbuckets(
  data_dir,
  output_dir,
  siteid,
  long.thres,
  data_type,
  pascphens,
  chart_review = FALSE
)
}
\arguments{
\item{data_dir}{directory of the 4CE tables}

\item{output_dir}{directory for saving the results}

\item{siteid}{label specifying the site}

\item{long.thres}{long" stage threshold}

\item{data_type}{string with the type of 4CE data}

\item{pascphens}{list of pasc phenotype to study}

\item{chart_review}{set TRUE if you want data for chart review}
}
\description{
analyzing buckets of MLHO and core features

analyzing buckets of MLHO and core features
}
