% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phen_distribution.R
\name{phen_distribution}
\alias{phen_distribution}
\title{Phenotype distribution}
\usage{
phen_distribution(
  PatientObservations,
  PatientSummary,
  rules,
  stratified_by = NULL
)
}
\arguments{
\item{PatientObservations}{2.1 PatientObservations table}

\item{PatientSummary}{2.1 PatientSummary table}

\item{rules}{table containing the phenotype definitions}

\item{stratified_by}{label for the stratification of the bar plot}
}
\value{
two plots
}
\description{
Build a ridgeline plot and a box plot of the distribution of the phenotypes
over the dataset
}
