% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUN.R
\name{RUN}
\alias{RUN}
\title{RUN PASC analysis}
\usage{
RUN(
  data_dir,
  output_dir,
  siteid,
  long.thres,
  long.perc = 0.03,
  MSMR.sparsity,
  data_type
)
}
\arguments{
\item{data_dir}{directory of the 4CE tables}

\item{output_dir}{directory of the output report. If NULL, the result is saved in 'getProjectOutput' directory}

\item{siteid}{label specifying the site}

\item{long.thres}{"long" stage threshold}

\item{long.perc}{percentage criterion to choose which phenotypes are interesting}

\item{MSMR.sparsity}{MSMR lite parameter}
}
\description{
RUN PASC analysis
}
