% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAnalysis_distribution.R
\name{runAnalysis_distribution}
\alias{runAnalysis_distribution}
\title{run distribution analysis}
\usage{
runAnalysis_distribution(data_dir, siteid, long.thres)
}
\arguments{
\item{data_dir}{directory of the 4CE tables}

\item{siteid}{label specifying the site}

\item{long.thres}{"long" stage threshold}
}
\value{
phenotype data and count tables
}
\description{
run distribution analysis
}
