% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAnalysis_MLHO.R
\name{runAnalysis_MLHO}
\alias{runAnalysis_MLHO}
\title{run MLHO analysis}
\usage{
runAnalysis_MLHO(data_dir, siteid, long.thres, phenotype, MSMR.sparsity)
}
\arguments{
\item{data_dir}{directory of the 4CE tables}

\item{siteid}{label specifying the site}

\item{long.thres}{"long" stage threshold}

\item{phenotype}{phenotype of interest}

\item{MSMR.sparsity}{MSMR lite parameter}
}
\value{
phenotyping output and features prevalance
}
\description{
run MLHO analysis
}
