% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phen_distribution.R
\name{phen_distribution}
\alias{phen_distribution}
\title{Phenotype distribution}
\usage{
phen_distribution(PatientObservations, PatientSummary, rules, siteid)
}
\arguments{
\item{PatientObservations}{2.1 PatientObservations table}

\item{PatientSummary}{2.1 PatientSummary table}

\item{rules}{table containing the phenotype definitions}

\item{siteid}{label specifying the site}
}
\value{
count tables
}
\description{
Phenotype distribution
}
