% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAnalysis_distribution.R
\name{runAnalysis_distribution}
\alias{runAnalysis_distribution}
\title{run distribution analysis}
\usage{
runAnalysis_distribution(data_dir, output_dir = NULL, siteid)
}
\arguments{
\item{data_dir}{directory of the 4CE tables}

\item{output_dir}{directory of the analysis results. If NULL results are saved in the \code{getProjectOutputDirectory()} directory}

\item{siteid}{label specifying the site}
}
\value{
NULL.
}
\description{
run distribution analysis
}
